+++ packages/native/src/components/NativeValueEditor.tsx
@@ -44,7 +44,7 @@ export const NativeValueEditor = ({
     ]
   );
 
-  const { valArray, betweenValueHandler } = useValueEditor({
+  const { valueAsArray, multiValueHandler } = useValueEditor({
     skipHook,
     handleOnChange,
     inputType,
@@ -77,9 +77,9 @@ export const NativeValueEditor = ({
             style={styles.value}
             keyboardType={keyboardType}
             placeholder={placeHolderText}
-            value={valArray[i] ?? ''}
+            value={valueAsArray[i] ?? ''}
             // TODO: disabled={disabled}
-            onChangeText={v => betweenValueHandler(v, i)}
+            onChangeText={v => multiValueHandler(v, i)}
           />
         );
       }
@@ -87,10 +87,10 @@ export const NativeValueEditor = ({
         <SelectorComponent
           key={key}
           {...props}
-          handleOnChange={v => betweenValueHandler(v, i)}
+          handleOnChange={v => multiValueHandler(v, i)}
           className={className}
           disabled={disabled}
-          value={valArray[i] ?? getFirstOption(values)}
+          value={valueAsArray[i] ?? getFirstOption(values)}
           options={values}
           listsAsArrays={listsAsArrays}
         />