+++ packages/react-querybuilder/src/components/ValueEditor.tsx
@@ -23,7 +23,7 @@ export const ValueEditor = ({
   selectorComponent: SelectorComponent = ValueSelector,
   ...props
 }: ValueEditorProps) => {
-  const { valArray, betweenValueHandler } = useValueEditor({
+  const { valueAsArray, multiValueHandler } = useValueEditor({
     skipHook,
     handleOnChange,
     inputType,
@@ -53,10 +53,10 @@ export const ValueEditor = ({
             key={key}
             type={inputTypeCoerced}
             placeholder={placeHolderText}
-            value={valArray[i] ?? ''}
+            value={valueAsArray[i] ?? ''}
             className={standardClassnames.valueListItem}
             disabled={disabled}
-            onChange={e => betweenValueHandler(e.target.value, i)}
+            onChange={e => multiValueHandler(e.target.value, i)}
           />
         );
       }
@@ -65,9 +65,9 @@ export const ValueEditor = ({
           key={key}
           {...props}
           className={standardClassnames.valueListItem}
-          handleOnChange={v => betweenValueHandler(v, i)}
+          handleOnChange={v => multiValueHandler(v, i)}
           disabled={disabled}
-          value={valArray[i] ?? getFirstOption(values)}
+          value={valueAsArray[i] ?? getFirstOption(values)}
           options={values}
           listsAsArrays={listsAsArrays}
         />